Refactored some shared functions into common.go
